[TCR] flatten nested conditionals
Now that all of the if expressions have the same simple structures, I recorded a
Vim macro to automate collapsing a three-line if..end expression into its
parent.  Then I started from the bottom of the file, searched backward for the
string 'if', and replayed the macro until everything was collapsed to the top
level.